requirements.txt
flask
tensorflow==2.19.0
pillow
numpy
gunicorn
